requirements.txt
spotipy
scikit-learn
numpy
python-dotenv
lyricsgenius
pandas